add WAN_WAIT

+++ src/deptran/paxos_worker.h
@@ -17,9 +17,9 @@ namespace janus {
   class BulkPaxosCmd;
 
 	inline void read_log(const char* log, int length, const char* custom){
-		unsigned long long int cid = 0;
-		memcpy(&cid, log, sizeof(unsigned long long int));
-		Log_info("commit id %lld and length %d from %s", cid, length, custom);
+		uint32_t cid = 0;
+		memcpy(&cid, log, sizeof(uint32_t));
+		Log_info("commit id %ld and length %d from %s", cid, length, custom);
 	}
 
 	inline size_t track_write(int fd, const void* p, size_t len, int offset){
@@ -593,7 +593,7 @@ private:
   bool noops_received=false;
   std::function<void(const char*, int)> callback_ = nullptr;
   std::function<void(const char*&, int, int)> callback_par_id_ = nullptr;
-  std::function<std::vector<uint64_t>(const char*&, int, int, int, std::queue<std::tuple<std::vector<uint64_t>, int, int, int, const char *>> &)> callback_par_id_return_ = nullptr;
+  std::function<std::vector<uint32_t>(const char*&, int, int, int, std::queue<std::tuple<std::vector<uint32_t>, int, int, int, const char *>> &)> callback_par_id_return_ = nullptr;
   vector<Coordinator*> created_coordinators_{};
   vector<shared_ptr<Coordinator>> created_coordinators_shrd{};
   struct timeval t1;
@@ -625,7 +625,7 @@ public:
   base::ThreadPool* hb_thread_pool_g = nullptr;
 
   Config::SiteInfo* site_info_ = nullptr;
-  std::queue<std::tuple<std::vector<uint64_t>, int, int, int, const char *>> un_replay_logs_ ;  // latest_commit_id, status, len, log
+  std::queue<std::tuple<std::vector<uint32_t>, int, int, int, const char *>> un_replay_logs_ ;  // latest_commit_id, status, len, log
   Frame* rep_frame_ = nullptr;
   TxLogServer* rep_sched_ = nullptr;
   Communicator* rep_commo_ = nullptr;
@@ -677,7 +677,7 @@ public:
   void Submit(const char*, int, uint32_t);
   void register_apply_callback(std::function<void(const char*, int)>);
   void register_apply_callback_par_id(std::function<void(const char*&, int, int)>);
-  void register_apply_callback_par_id_return(std::function<vector<uint64_t>(const char*&, int, int, int, std::queue<std::tuple<std::vector<uint64_t>, int, int, int, const char *>> &)>);
+  void register_apply_callback_par_id_return(std::function<vector<uint32_t>(const char*&, int, int, int, std::queue<std::tuple<std::vector<uint32_t>, int, int, int, const char *>> &)>);
   rrr::PollMgr * GetPollMgr(){
       return svr_poll_mgr_;
   }
